feature/event-details #fixed dark theme
